% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMathR-package.R
\docType{package}
\name{BioMathR-package}
\alias{BioMathR}
\alias{BioMathR-package}
\title{BioMathR: Utility functions used at BioMath}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The R package `BioMathR` provides a collection of utility and wrapper functions, all of which are often used at [BioMath](https://www.biomath.de/).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://schmidtpaul.github.io/BioMathR/}
  \item \url{https://github.com/SchmidtPaul/BioMathR}
  \item Report bugs at \url{https://github.com/SchmidtPaul/BioMathR/issues}
}

}
\author{
\strong{Maintainer}: Paul Schmidt \email{paul.schmidt@biomath.de}

Other contributors:
\itemize{
  \item BioMath [funder]
}

}
\keyword{internal}
