% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wb.R
\name{create_wb}
\alias{create_wb}
\title{Create a Workbook object}
\usage{
create_wb(
  font_size = 10,
  font_name = "Arial",
  infosheet = TRUE,
  infosheet_label = "BioMath GmbH",
  ...,
  fontSize = deprecated(),
  fontName = deprecated(),
  infosheetlabel = deprecated()
)
}
\arguments{
\item{font_size}{font size}

\item{font_name}{Name of a font}

\item{infosheet}{Should an info sheet be created?}

\item{infosheet_label}{Label for the info sheet}

\item{...}{Other arguments passed to \code{openxlsx2::wb_workbook()}}

\item{fontSize}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{font_size}.}

\item{fontName}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{font_name}.}

\item{infosheetlabel}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{infosheet_label}.}
}
\value{
Workbook object (see \{openxlsx2\})
}
\description{
This function is a wrapper for \code{openxlsx2::wb_workbook()}, but also sets the base font and adds an info sheet including e.g. the time it was created.
}
\examples{
wb1 <- create_wb()

wb2 <- create_wb(infosheet = FALSE)
}
