% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Compute descriptive statistics}
\usage{
describe(data, yvars, ungroupafter = TRUE, lang = c("eng", "ger")[1])
}
\arguments{
\item{data}{a data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{yvars}{a character vector with the names of the columns to be described.}

\item{ungroupafter}{If \code{TRUE}, the results will be \code{ungroup()}-ed.}

\item{lang}{If \code{"eng"} (default) or  \code{"ger"} columns will be renamed accordingly.}
}
\value{
A tibble
}
\description{
This function computes common descriptive statistics of numeric variables for grouped or ungrouped data and converts it to a tidy formatted table. This function tried to copy \code{dlookr::describe()} because of \href{https://github.com/choonghyunryu/dlookr/issues/79}{this issue}.
}
\examples{
library(dplyr)
library(BioMathR)
PlantGrowth \%>\%
  group_by(group) \%>\%
  describe("weight")

PlantGrowth \%>\%
  mutate(weight2 = weight + 1) \%>\%
  group_by(group) \%>\%
  describe(c("weight", "weight2"))
}
