% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans_BM.R
\name{emmeans_BM}
\alias{emmeans_BM}
\title{Deprecated version of get_emmeans}
\usage{
emmeans_BM(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{get_emmeans}}.}
}
\description{
This function has been deprecated. Please use \code{\link{get_emmeans}} instead.
}
\seealso{
\code{\link{get_emmeans}}
}
