% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dlookrdescribe.R
\name{format_dlookrdescribe}
\alias{format_dlookrdescribe}
\title{Format dlookr::describe() output}
\usage{
format_dlookrdescribe(dlookr_describe_out, lang = c("ger", "eng")[1])
}
\arguments{
\item{dlookr_describe_out}{a table create via \code{dlookr::describe()}}

\item{lang}{language for the formatted result table and text}
}
\description{
Format dlookr::describe() output
}
