% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anova.R
\name{get_anova}
\alias{get_anova}
\alias{get_anova.lm}
\alias{get_anova.lmerMod}
\alias{get_anova.glmmTMB}
\title{Obtain analysis of variance table}
\usage{
get_anova(model, type, ...)

\method{get_anova}{lm}(model, type = c("I", "II", "III")[3], ...)

\method{get_anova}{lmerMod}(model, type = c("I", "II", "III")[3], ...)

\method{get_anova}{glmmTMB}(model, type = c("I", "II", "III")[3], ...)
}
\arguments{
\item{model}{a fitted model object. Supported classes: \code{"lm"}, \code{"lmerMod"} & \code{"glmmTMB"}.}

\item{type}{Type of ANOVA test: \code{"I"}, \code{"II"} or \code{"III"} (default).}

\item{...}{additional arguments to be passed to the specific anova function.}
}
\description{
This function obtains the ANOVA table for a model.
}
\details{
The function utilizes \code{car::Anova()} for all supported model classes. For models of class 'lm' and 'lmerMod', the F-test is employed. In contrast, for models of the 'glmmTMB' class, the Chi-Square test is used since the F-test is \href{https://github.com/glmmTMB/glmmTMB/blob/bd1932addbfb4edf2ce933675f5a8bf72abc0e7c/glmmTMB/R/Anova.R#L73C8-L73C8}{currently unavailable}. It's worth noting that only the 'lmerMod' models support (and default to) the Kenward-Roger method as a degrees of freedom method.
}
\seealso{
[car::Anova()],
  [lmerTest::lmer()],
  [glmmTMB::glmmTMB()]
}
