% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residual_plots.R
\name{get_residual_plots}
\alias{get_residual_plots}
\title{Get residual plots for model diagnostics}
\usage{
get_residual_plots(
  model,
  lang = c("eng", "ger")[1],
  col_dots = BioMathR::palette_getset("BioMath", setforggplot = FALSE)[["grey"]],
  col_line = BioMathR::palette_getset("BioMath", setforggplot = FALSE)[["green"]],
  aslist = FALSE
)
}
\arguments{
\item{model}{A model object.}

\item{lang}{Language for plots labels.}

\item{col_dots}{Color of the dots.}

\item{col_line}{Color of the lines.}

\item{aslist}{If \code{TRUE}, output is a list of plots. If \code{FALSE} (default), output is a combined plot.}
}
\description{
Get residual plots for model diagnostics
}
