% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_references.R
\name{get_unique_references}
\alias{get_unique_references}
\title{Get unique reference notations}
\usage{
get_unique_references(x)
}
\arguments{
\item{x}{a character vector with multiple strings/references}
}
\value{
a character vector
}
\description{
When multiple references (formatted as e.g. `Dawkins (2008)`) have an identical author (or authors) and publication year, it is common practice to include a lowercase letter after the year. This function does this for a vector containing multiple references.
}
\examples{
refs <- c("Dawkins (2008)", "Dawkins (2008)", "Stephenson (2008)")

BioMathR::get_unique_references(refs)
}
