% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Helper function: Negation/Opposite of `\%in\%`}
\usage{
a \%not_in\% b
}
\arguments{
\item{a}{vector or `NULL`: the values to be matched.}

\item{b}{vector or `NULL`: the values to be matched against.}
}
\value{
A logical vector, indicating if a match was
not located for each element of `a`. The values
are `TRUE` or `FALSE` and never `NA`.
}
\description{
This helper function checks if elements
of one vector are not found in another vector.
}
\examples{
c(1,2,3) \%not_in\% c(3,4,5,6,7)

library(dplyr)
PlantGrowth \%>\% filter(group \%not_in\% c("ctrl", "trt2"))
}
