% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_anovas.R
\name{join_anovas}
\alias{join_anovas}
\title{Join Multiple ANOVA Tables}
\usage{
join_anovas(anova_list, anova_names)
}
\arguments{
\item{anova_list}{A list of ANOVA tables.}

\item{anova_names}{A character vector specifying the names for each ANOVA in the list.}
}
\value{
A tibble with merged ANOVA results.
}
\description{
This function takes a list of ANOVA tables and joins them into a single tibble.
If an element of the ANOVA list is NULL, then an empty column is added.
}
