% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_getset.R
\name{palette_getset}
\alias{palette_getset}
\title{Get and set color palettes}
\usage{
palette_getset(palette, setforggplot = TRUE)
}
\arguments{
\item{palette}{Choose color palette \itemize{ \item{\code{"BioMath"} - \href{https://www.biomath.de/}{BioMath} corporate design (\href{https://coolors.co/f0a202-495057-00923f-201e50-e4572e}{see preview})} \item{\code{"UBA"} - Umweltbundesamt \href{https://www.umweltbundesamt.de/dokument/corporate-design-des-umweltbundesamtes}{corporate design} (\href{https://coolors.co/5ead35-007626-0b90d5-005f85-9d579a-622f63-ce1f5e-83053c-fabb00-d78400}{see preview})}}}

\item{setforggplot}{If \code{TRUE} (default), the color palette is set as default \code{ggplot2.discrete.fill} and \code{ggplot2.discrete.colour}.}
}
\value{
The color palette as a named vector
}
\description{
Get and set color palettes
}
\examples{
library(ggplot2)

p1 <- ggplot(data = PlantGrowth) +
  aes(y = weight, x = group, fill = group) +
  geom_bar(stat = "summary", fun = "mean")

p1

BioMathR::palette_getset("BioMath")

p1

p1 + BioMathR::theme_BioMath()
}
