% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{project_setup}
\alias{project_setup}
\title{Intitiate an R-project}
\usage{
project_setup()
}
\value{
Prints a success message if the directories and setup file are created successfully.
}
\description{
This function intitiates a new R-project with the typical folder structure
and a \code{00 setup.R} file.
}
\details{
The root directory "10 RCode" is created in the working directory, and its
sub-directories are "code", "data", "src", and "out". The function also
creates a setup file "00 setup.R" in the "code" sub-directory. This file
contains calls to load certain libraries and defines a few utility
functions and a color palette. If the 'here' package is not installed, the
function stops and asks the user to install it.
}
