% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_smart.R
\name{round_smart}
\alias{round_smart}
\title{Automatically round numbers}
\usage{
round_smart(x, signif_digits = 1, max_digits = 6)
}
\arguments{
\item{x}{A number of vector of numbers}

\item{signif_digits}{The number of significant digits (default: 1) like in \code{signif()}, but only applied to the part after the decimal separator.
\itemize{
\item \code{round_smart(1.0001234, signif_digits = 1)} returns \code{1.0001}
\item \code{round_smart(1.0001234, signif_digits = 2)} returns \code{1.00012}
}}

\item{max_digits}{The maximum number of digits (default: 6) to round to.
\itemize{
\item \code{round_smart(1.0000001234, max_digits = 6)} returns \code{1}
\item \code{round_smart(1.0000001234, max_digits = Inf)} returns \code{1.0000001}
}}
}
\description{
This function automatically rounds numbers so that the results have as few digits as possible, but als many as necessary. It is similar to \code{signif()}, but it never alters the part before the decimal separator and also allows a cutoff value for a maximum acceptable number of digits after the decimal separator.
}
\examples{
library(BioMathR)
library(dplyr)

before <- data.frame(
V1 = c(123456, 1234),
V2 = c(-123, -0.12345),
V3 = c(1.0012345, 0.1),
V4 = c(1.1, 0.0012345),
V5 = c(1.000000012345, 0),
V6 = c(NA, -5.0018)
)

mutate(before, across(everything(), ~round_smart(.)))
mutate(before, across(everything(), ~round_smart(., signif_digits = 3)))
mutate(before, across(everything(), ~round_smart(., signif_digits = 3, max_digits = Inf)))
}
