% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wb.R
\name{save_wb}
\alias{save_wb}
\title{Save a Workbook object to Excel file}
\usage{
save_wb(
  wb,
  file = here::here("untitled.xlsx"),
  overwrite = FALSE,
  open_after_save = TRUE
)
}
\arguments{
\item{wb}{A Workbook object to write to file}

\item{file}{A character string naming an xlsx file}

\item{overwrite}{If \code{TRUE}, overwrites the previous excel file}

\item{open_after_save}{If \code{TRUE}, opens the excel file after it was created}
}
\value{
The file path, invisibly
}
\description{
This function is a wrapper for \code{openxlsx2::wb_save()}, but can also open the Excel file immediately after saving. File opening is cross-platform compatible (Windows, macOS, and Linux). If the first sheet is the \code{"info"} sheet of \code{create_wb()} and there is a second sheet, that second sheet is the active and selected one when the file is opened.
}
