% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse_with.R
\name{str_collapse_with}
\alias{str_collapse_with}
\title{Collapse a Vector of Strings with Custom Separators}
\usage{
str_collapse_with(
  strings,
  sep = ", ",
  final_sep = " and ",
  wrap_each_in = NULL,
  error_when_NULL = TRUE,
  error_when_NA = TRUE,
  na_replace = "-"
)
}
\arguments{
\item{strings}{A character vector to be collapsed.}

\item{sep}{A string separator to use between elements, default is ", ".}

\item{final_sep}{A string separator to use before the last element, default is " and ".}

\item{wrap_each_in}{A single character or a string of two characters used to wrap each string element. If two characters are given, the first is used as the prefix and the second as the suffix.}

\item{error_when_NULL}{Logical; if TRUE (default), the function throws an error when \code{strings} is NULL or empty.}

\item{error_when_NA}{Logical; if TRUE (default), the function throws an error when \code{strings} contains NA or empty strings.}

\item{na_replace}{The string to replace NA or empty strings with, if \code{error_when_NA} is FALSE.}
}
\value{
A single collapsed string. If \code{strings} is NULL or empty and \code{error_when_NULL} is FALSE, NULL is returned.
}
\description{
This function takes a vector of strings and collapses it into a single string,
using specified separators. By default, an error is thrown if the vector is NULL,
empty, contains NA values, or empty strings, but these behaviors can be modified.
}
\examples{
for(i in 1:4){
  print(str_collapse_with(letters[1:i]))
}

str_collapse_with(levels(PlantGrowth$group), final_sep = " und ", wrap_each_in = '`')
str_collapse_with(letters[1:5], sep = " & ", final_sep = " as well as ", wrap_each_in = '{}')

}
