% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_unGer.R
\name{str_unGer}
\alias{str_unGer}
\title{Replace German special characters in string}
\usage{
str_unGer(string)
}
\arguments{
\item{string}{a character string possibly including Ä, ä, Ö, ö, Ü, ü or ß}
}
\value{
a string
}
\description{
This function replaces Umlaute (Ä, ä, Ö, ö, Ü, ü) and the letter ß in a string.
}
\examples{
BioMathR::str_unGer("Über grüßen; Öfter größer; Änderung ändern; Ä, ä, Ö, ö, Ü, ü or ß.")
}
