% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wrap_br.R
\name{str_wrap_br}
\alias{str_wrap_br}
\title{`str_wrap()` with <br> instead}
\usage{
str_wrap_br(string, whitespace_only = FALSE, ...)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{whitespace_only}{If `TRUE` wrapping will only occur at whitespace. If `FALSE` (the default), can break on any non-word character (e.g. /, -). Note that this default is different from the underlying `stringr::str_wrap()`.}

\item{...}{other arguments passed to `stringr::str_wrap()`}
}
\description{
`str_wrap()` with <br> instead
}
