% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_UBA.R
\name{theme_UBA}
\alias{theme_UBA}
\title{A \{ggplot2\} theme based on the \href{https://www.umweltbundesamt.de/dokument/corporate-design-des-umweltbundesamtes}{Umweltbundesamt corporate design}.}
\usage{
theme_UBA(
  base_size = 11,
  base_family = "",
  base_color = "black",
  axis_title_bold = TRUE,
  axis_title_just = "cc",
  ticks = FALSE,
  grid_x = TRUE,
  grid_y = TRUE,
  grid_color = "black",
  grid_linetype = "solid",
  facette_distance = 1,
  facette_box = TRUE,
  facette_box_color = "black",
  title_size = 12,
  title_margin = margin(b = 10),
  subtitle_size = 9,
  subtitle_margin = margin(b = 10),
  axistext_size = 10,
  is_subplot = FALSE,
  legend_custom_bottom = FALSE
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family. Should be Calibri, but is not set as default.}

\item{base_color}{Base color for text and lines}

\item{axis_title_bold}{Should axis titles be bold?}

\item{axis_title_just}{Justification of axis titles. \code{rt} translates to "right" (x-axis) "top" (y-axis).}

\item{ticks}{Should ticks be drawn on the axes?}

\item{grid_x}{Should there be major grid lines for the x-axis?}

\item{grid_y}{Should there be major grid lines for the y-axis?}

\item{grid_color}{Color of the grid lines}

\item{grid_linetype}{Linetype of the grid lines}

\item{facette_distance}{Distance between facettes}

\item{facette_box}{Should a rectangle be drawn around each facette?}

\item{facette_box_color}{Color of the rectangle around each facette}

\item{title_size}{Title font size, given in pts.}

\item{title_margin}{Margin around tile}

\item{subtitle_size}{Subtitle font size, given in pts.}

\item{subtitle_margin}{Margin around subtitle}

\item{axistext_size}{Axis text font size, given in pts.}

\item{is_subplot}{If \code{TRUE}, font sizes and margins of title, subtitle and axis text are decreased. This is useful if multiple plots are to be combined.}

\item{legend_custom_bottom}{If \code{TRUE}, the legend is formatted in a specific way at the bottom of the plot}
}
\description{
A \{ggplot2\} theme based on the \href{https://www.umweltbundesamt.de/dokument/corporate-design-des-umweltbundesamtes}{Umweltbundesamt corporate design}.
}
\examples{
library(ggplot2)

p1 <- ggplot(data = PlantGrowth) +
  aes(y = weight, x = group, fill = group) +
  geom_bar(stat = "summary", fun = "mean")

BioMathR::palette_getset("UBA")

p1

p1 + BioMathR::theme_UBA()
}
