% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_reg.R
\name{tidy_reg}
\alias{tidy_reg}
\alias{tidy_reg.lm}
\alias{tidy_reg.openair}
\title{Summary of linear regression (for a time series)}
\usage{
tidy_reg(x, reg_dat, n_preds = 10, ...)

\method{tidy_reg}{lm}(x, reg_dat, n_preds = 10, ..., conf.level = 0.95)

\method{tidy_reg}{openair}(x, reg_dat, n_preds = 10, ..., adjust_x0 = TRUE)
}
\arguments{
\item{x}{A regression object created via \code{stats::lm()} or \code{openair::TheilSen()}}

\item{reg_dat}{The exact same dataset that was also provided when creating \code{x} above}

\item{n_preds}{Number of evenly spread out points on the x-axis for which predictions (including confidence interval) should be made}

\item{...}{Other arguments passed to \code{broom::tidy()} from \code{lm} models.}

\item{conf.level}{The confidence level to use for the confidence interval}

\item{adjust_x0}{If \code{TRUE} (default) the estimate and confidence limits for the intercept are adjusted so that it is no longer at {openair}'s default x = 1970, but instead at the minimum x present in the data.}
}
\description{
Summary of linear regression (for a time series)
}
