% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ascii.R
\name{write_ascii}
\alias{write_ascii}
\title{Write a Data Frame to an ASCII Text File}
\usage{
write_ascii(data, path, dec = ".")
}
\arguments{
\item{data}{A data frame that will be written to a text file.}

\item{path}{The file path where the text file will be written.}

\item{dec}{The character to be used as decimal point. Default is ".".}
}
\value{
NULL. The function is invoked for its side effect, which is writing the data frame
to a text file.
}
\description{
This function takes a data frame, a file path, and a decimal point character as input,
and writes the data frame to the file as a tab-separated ASCII text file,
with no quotes around the values, no row names. The decimal point representation
is by default a dot, but can be changed by the user.
}
