% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add a worksheet to a Workbook object}
\usage{
add_sheet(
  wb,
  data,
  sheetName = NULL,
  colWidth = "auto",
  colWidthMin = 5,
  colWidthMax = 25,
  gridLines = FALSE,
  freezefirstRow = TRUE,
  freezefirstCol = FALSE,
  addFilters = TRUE,
  rowheight = 15,
  textwrap = TRUE,
  dateformat = "yyyy-mm-dd",
  datetimeformat = "yyyy-mm-dd hh:mm:ss"
)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{data}{Object to be written. For classes supported look at the examples in {openxlsx}.}

\item{sheetName}{The name of the worksheet that is to be created and have `data` written into. Defaults to the `data` object name if left empty.}

\item{colWidth}{For all columns: column width. Can either be Excel column width units or "auto" for automatic sizing}

\item{colWidthMin}{For all columns: minimum column width}

\item{colWidthMax}{For all columns: maximum column width}

\item{gridLines}{If `FALSE`, the worksheet grid lines will be hidden}

\item{freezefirstRow}{If `TRUE`, freezes the first row}

\item{freezefirstCol}{If `TRUE`, freezes the first row}

\item{addFilters}{If `TRUE`, adds filters to worksheet columns}

\item{rowheight}{Row height in Excel row height units}

\item{textwrap}{If `TRUE`, all cells are formatted via text wrap so their content never spills into adjacent empty cells}

\item{dateformat}{If not `"none"` (default is `"yyyy-mm-dd"`), all date-formatted columns in `data` are explicitly formatted with `numFmt = "yyyy-mm-dd"`}

\item{datetimeformat}{If not `"none"` (default is `"yyyy-mm-dd hh:mm:ss"`), all datetime-formatted columns in `data` are explicitly formatted with `numFmt = "yyyy-mm-dd hh:mm:ss"`}
}
\description{
This function is a wrapper for `openxlsx::addWorksheet()` and `openxlsx::writeData()`, but also does some extra formatting via more {openxlsx} functions.
}
\examples{
library(BioMathR)
wb1 <- create_wb()
wb1 <- add_sheet(wb = wb1, data = mtcars)

library(dplyr, warn.conflicts = FALSE)
wb2 <- create_wb() \%>\%
   add_sheet(mtcars, "The cars sheet")
}
