% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add a worksheet to a Workbook object}
\usage{
add_sheet(
  wb,
  data,
  sheetName = NULL,
  colWidth = "auto",
  colWidthMin = 5,
  colWidthMax = 25,
  gridLines = FALSE,
  freezefirstRow = TRUE,
  freezefirstCol = FALSE,
  addFilters = TRUE
)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet}

\item{data}{Object to be written. For classes supported look at the examples in {openxlsx}.}

\item{sheetName}{The name of the worksheet that is to be created and have \code{data} written into. Defaults to the \code{data} object name if left empty.}

\item{colWidth}{For all columns: column width. Can either be Excel column width units or "auto" for automatic sizing.}

\item{colWidthMin}{For all columns: minimum column width}

\item{colWidthMax}{For all columns: maximum column width}

\item{gridLines}{If \code{FALSE}, the worksheet grid lines will be hidden.}

\item{freezefirstRow}{If \code{TRUE}, freezes the first row}

\item{freezefirstCol}{If \code{TRUE}, freezes the first row}

\item{addFilters}{If \code{TRUE}, adds filters to worksheet columns}
}
\description{
This function is a wrapper for \code{openxlsx::addWorksheet()} and \code{openxlsx::writeData()}, but also does some extra formatting via more {openxlsx} functions.
}
\examples{
library(BioMathR)
wb1 <- create_wb()
wb1 <- add_sheet(wb = wb1, data = mtcars)

library(dplyr, warn.conflicts = FALSE)
wb2 <- create_wb() \%>\%
   add_sheet(mtcars, "The cars sheet")
}
