% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMathR-package.R
\docType{package}
\name{BioMathR-package}
\alias{BioMathR}
\alias{BioMathR-package}
\title{BioMathR: Utility functions used at BioMath}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The R package `BioMathR` provides a collection of utility and wrapper functions, {ggplot2} themes and color palettes, all of which are often used at [BioMath](https://www.biomath.de/).
}
\author{
\strong{Maintainer}: Paul Schmidt \email{paul.schmidt@biomath.de}

Other contributors:
\itemize{
  \item BioMath [funder]
}

}
\keyword{internal}
