% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wb.R
\name{create_wb}
\alias{create_wb}
\title{Create a Workbook object}
\usage{
create_wb(
  fontSize = 10,
  fontName = "Arial",
  infosheet = TRUE,
  infosheetlabel = "BioMath GmbH",
  ...
)
}
\arguments{
\item{fontSize}{font size}

\item{fontName}{Name of a font}

\item{infosheet}{Should an info sheet be created?}

\item{infosheetlabel}{Label for the info sheet}

\item{...}{Other arguments passed to `openxlsx::createWorkbook()`}
}
\value{
Workbook object (see {openxlsx})
}
\description{
This function is a wrapper for `openxlsx::createWorkbook()`, but also adds an info sheet including e.g. the time it was created.
}
\examples{
wb1 <- create_wb()

wb2 <- create_wb(infosheet = FALSE)
}
