% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_tabs.R
\name{desc_tabs}
\alias{desc_tabs}
\title{Get all descriptive tables}
\usage{
desc_tabs(
  data,
  yvars,
  groupvars,
  lang = c("eng", "ger")[1],
  xlsx_path = here::here("desctabs.xlsx"),
  xlsx_overwrite = FALSE,
  xlsx_open = FALSE,
  xlsx_data_sheet = TRUE
)
}
\arguments{
\item{data}{Data object.}

\item{yvars}{Column names of variables for which descriptive measures should be calculated.}

\item{groupvars}{Column names of variables for which grouping should be done.}

\item{lang}{Language for the result table column names.}

\item{xlsx_path}{Path to where the excel file should be saved.}

\item{xlsx_overwrite}{If `TRUE`, overwrites the previous excel file.}

\item{xlsx_open}{If `TRUE`, opens the excel file after it was created.}

\item{xlsx_data_sheet}{If `TRUE`, adds a excel sheet with the underyling data.}
}
\description{
Get all descriptive tables
}
