% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Compute descriptive statistics}
\usage{
describe(data, yvars, ungroupafter = TRUE, lang = c("eng", "ger")[1])
}
\arguments{
\item{data}{a data.frame or a `\link{tbl_df}` or a `\link{grouped_df}`.}

\item{yvars}{a character vector with the names of the columns to be described.}

\item{ungroupafter}{If `TRUE`, the results will be `ungroup()`-ed.}

\item{lang}{Language for table names.}
}
\value{
A tibble
}
\description{
This function computes common descriptive statistics of numeric variables for grouped or ungrouped data and converts it to a tidy formatted table. This function tried to copy `dlookr::describe()` because of \href{https://github.com/choonghyunryu/dlookr/issues/79}{this issue}.
}
\examples{
library(dplyr)
library(BioMathR)
PlantGrowth \%>\%
  group_by(group) \%>\%
  describe("weight", lang = "ger")

PlantGrowth \%>\%
  mutate(
    group2 = as.factor(rep(c("A", "A", "B",  "B", "B"), 6)),
    weight2 = weight * 2,
    weight3 = rep(c(3, NA), 15)
  ) \%>\%
  group_by(group, group2) \%>\%
  describe(c("weight", "weight2", "weight3"))
}
