% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Compute descriptive statistics}
\usage{
describe(
  data,
  yvars,
  lang = c("eng", "ger")[1],
  ungroupafter = TRUE,
  digits = "round_smart",
  addstats = character(),
  ...
)
}
\arguments{
\item{data}{A data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{yvars}{A character vector with the names of the columns to be described.}

\item{lang}{Language for table names.}

\item{ungroupafter}{If \code{TRUE}, the results will be \code{ungroup()}-ed.}

\item{digits}{Number of digits all numeric columns are rounded to. The default is actually \code{"round_smart"} which applies \code{BioMathR::round_smart()} to each numeric column individually.}

\item{addstats}{Character vector with additional statistics to be calculated besides the default. Possible entries: c("Skewness").}

\item{...}{Other arguments passed to \code{BioMathR::round_smart()}}
}
\value{
A tibble
}
\description{
This function computes common descriptive statistics of numeric variables for grouped or ungrouped data and converts it to a tidy formatted table. This function tried to copy \code{dlookr::describe()} because of \href{https://github.com/choonghyunryu/dlookr/issues/79}{this issue}.
}
\examples{
library(dplyr)
library(BioMathR)
PlantGrowth \%>\%
  group_by(group) \%>\%
  describe("weight")

PlantGrowth \%>\%
  mutate(
    group2 = as.factor(rep(c("A", "A", "B",  "B", "B"), 6)),
    weight2 = weight * 2,
    weight3 = rep(c(3, NA), 15)
  ) \%>\%
  group_by(group, group2) \%>\%
  describe(c("weight", "weight2", "weight3"), lang = "ger") \%>\%
  docx_tab()
}
