% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplot_across.R
\name{desplot_across}
\alias{desplot_across}
\title{Wrapper function for \code{desplot::desplot()} across multiple columns}
\usage{
desplot_across(
  data,
  vars,
  formright = "col + row",
  lang = c("eng", "ger")[1],
  title = c("none", "short", "long")[3],
  flip = TRUE,
  out1.gpar = list(col = "black", lwd = 1, lty = "dotted"),
  out2.gpar = list(col = "black", lwd = 1, lty = "dotted"),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Vector with variables/column names for which a desplot should be created.}

\item{formright}{A formula like \code{x*y|location}, i.e. the right-hand side of the formula \code{yield~x*y|location} that would usually be passed to \code{desplot::desplot(form = ...)}. Note that \code{x} and \code{y} are numeric and the default is \code{"col + row"}.}

\item{lang}{Language for plots labels.}

\item{title}{Can either be \code{"none"}, \code{"short"} or \code{"long"}. For the respective \code{var}, it gives information about the number of levels and their respective frequency in the data.}

\item{flip}{see \code{desplot::desplot()} documentation - set to opinionated default here}

\item{out1.gpar}{see \code{desplot::desplot()} documentation - set to opinionated default here}

\item{out2.gpar}{see \code{desplot::desplot()} documentation - set to opinionated default here}

\item{...}{Other arguments passed to \code{desplot::desplot()}.}
}
\value{
A list of desplots.
}
\description{
The goal of this function is to allow the user a quick and easy glance at the experimental layout of a trial by creating multiple \code{desplot}s at once - one for each variable that is provided (see example below).
}
\examples{
library(BioMathR)

dps <- desplot_across(data = agridat::yates.oats,
                      vars = c("nitro", "gen", "block"),
                      cex = 1)

dps$nitro
dps$gen
dps$block
}
