% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplot_across.R
\name{desplot_across}
\alias{desplot_across}
\title{Wrapper function for `desplot::desplot()` across multiple columns}
\usage{
desplot_across(
  data,
  vars,
  formright = "col + row",
  lang = c("eng", "ger")[1],
  title = c("none", "short", "long")[3],
  flip = TRUE,
  out1.gpar = list(col = "black", lwd = 1, lty = "dotted"),
  out2.gpar = list(col = "black", lwd = 1, lty = "dotted"),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{Vector with variables/column names for which a desplot should be created.}

\item{formright}{A formula like `x*y|location`, i.e. the right-hand side of the formula `yield~x*y|location` that would usually be passed to `desplot::desplot(form = ...)`. Note that `x` and `y` are numeric and the default is `"col + row"`.}

\item{lang}{Language for plots labels.}

\item{title}{Can either be `"none"`, `"short"` or `"long"`. For the respective `var`, it gives information about the number of levels and their respective frequency in the data.}

\item{flip}{see `desplot::desplot()` documentation - set to opinionated default here}

\item{out1.gpar}{see `desplot::desplot()` documentation - set to opinionated default here}

\item{out2.gpar}{see `desplot::desplot()` documentation - set to opinionated default here}

\item{...}{Other arguments passed to `desplot::desplot()`.}
}
\value{
A list of desplots.
}
\description{
The goal of this function is to allow the user a quick and easy glance at the experimental layout of a trial by creating multiple `desplot`s at once - one for each variable that is provided (see example below).
}
\examples{
library(BioMathR)

dps <- desplot_across(data = agridat::yates.oats,
                      vars = c("nitro", "gen", "block"),
                      cex = 1)

dps$nitro
dps$gen
dps$block
}
