% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_tab.R
\name{docx_tab}
\alias{docx_tab}
\title{Auto-format tables for printing in MS-Word documents}
\usage{
docx_tab(
  x,
  lang = c("eng", "ger")[1],
  digits = 1,
  pvalform = "p.value",
  asft = TRUE
)
}
\arguments{
\item{x}{Table to be formatted}

\item{lang}{Language for column names.}

\item{digits}{Number of digits all numeric columns are rounded to}

\item{pvalform}{Names of columns are that formatted via \code{BioMathR::format_p()}. Can be set to \code{NULL}. The default is \code{"p.value"}, but note that this function first unifies multiple column names such as \code{"Pr(>F)"} or \code{"P(>|Chi|)"} into \code{"p.value"}.}

\item{asft}{If \code{TRUE}, output is formatted as flextable}
}
\description{
Auto-format tables for printing in MS-Word documents
}
