% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_tab.R
\name{docx_tab}
\alias{docx_tab}
\title{Auto-format tables for printing in MS-Word documents}
\usage{
docx_tab(
  x,
  lang = c("eng", "ger")[1],
  digits = 1,
  autopvalform = TRUE,
  asft = TRUE
)
}
\arguments{
\item{x}{table to be formatted}

\item{lang}{Language for column names.}

\item{digits}{number of digits all numeric columns are rounded to}

\item{autopvalform}{if \code{TRUE}, p-value columns are formatted via \code{BioMathR::format_p()}}

\item{asft}{If \code{TRUE}, output is formatted as flextable}
}
\description{
Auto-format tables for printing in MS-Word documents
}
