% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans_BM.R
\name{emmeans_BM}
\alias{emmeans_BM}
\title{Get adjusted means with compact letter display}
\usage{
emmeans_BM(
  model,
  specs_string,
  lmer.df = c("Asymptotic", "Satterthwaite", "Kenward-Roger")[3],
  adjust = c("none", "tukey", "bonferroni")[1],
  lang = c("eng", "ger")[1],
  num_accuracy = 0.01,
  sort = c("none", "asc", "desc")[1]
)
}
\arguments{
\item{model}{a linear model object. See the \href{https://cran.r-project.org/web/packages/emmeans/vignettes/models.html}{list of supported model classes}}

\item{specs_string}{a character vector specifying the names of the predictors over which emmeans are desired}

\item{lmer.df}{degrees of freedom method. See the \href{https://cran.r-project.org/web/packages/emmeans/vignettes/sophisticated.html#lmerOpts}{list of available methods}}

\item{adjust}{\href{https://cran.r-project.org/web/packages/emmeans/vignettes/confidence-intervals.html#adjust}{multiplicity adjustment}. Follow \href{https://www.rdocumentation.org/packages/emmeans/versions/1.6.2-1/topics/summary.emmGrid}{this link} and scroll down o to the “P-value adjustments” heading within the “summary.emmGrid” section for more details on e.g. Fisher’s LSD test, Tukey-test, Bonferroni adjustment etc.}

\item{lang}{Language for table names and text.}

\item{num_accuracy}{format the number of digits shown after the decimal separator}

\item{sort}{sort the levels of the predictors given in `specs_string` in the output}
}
\description{
This function is a wrapper for `emmeans::emmeans()` and `multcomp::cld()` (\href{https://schmidtpaul.github.io/dsfair_quarto/summaryarticles/compactletterdisplay.html}{details here}). Based on a given linear model object, it computes estimated marginal means (a.k.a. least-squares means, adjusted means, modelbased means) for specified factors or factor combinations as well as comparisons or contrasts among them. Moreover, it adds the compact letter display to the comparisons/contrasts. Finally, its output is a list with six elements: for the means and their diffs, respectively, (i) a tibble with the results, (ii) the same tibble but in a print-ready format and (iii) an info text summarizing the results.
}
\details{
\href{https://github.com/rvlenth/emmeans/issues/235}{Here} is a discussion on the speed of calculating confidence intervals with different degrees of freedom methods.
}
