% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varcomp.R
\name{get_varcomp}
\alias{get_varcomp}
\alias{get_varcomp.merMod}
\alias{get_varcomp.glmmTMB}
\title{Extract variance components}
\usage{
get_varcomp(model, digits = 3)

\method{get_varcomp}{merMod}(model, digits = 3)

\method{get_varcomp}{glmmTMB}(model, digits = 3)
}
\arguments{
\item{model}{a fitted model object}

\item{digits}{Rounding. Default is 3.}
}
\description{
This function extracts the estimated variance components of the random-effects terms in a linear mixed model. It is very heavily based on \href{https://m-clark.github.io/mixedup/reference/extract_vc.html}{`mixedup::extract_vc()`}.
}
\seealso{
[lme4::VarCorr()],
  [glmmTMB::VarCorr()],
  [nlme::VarCorr()],
  [mixedup::extract_vc()]
}
