% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_export.R
\name{gg_export}
\alias{gg_export}
\title{Conveniently export (gg)plots}
\usage{
gg_export(
  plot_obj = NULL,
  folder_path = here::here(),
  file_name = "temp",
  width_cm = 16,
  height_cm = 10,
  pdf = "none",
  png = "create",
  svg = "none",
  png_dpi = 300,
  bg = "white",
  png_from_pdf = FALSE,
  svg_device = "ggsave"
)
}
\arguments{
\item{plot_obj}{Plot object to save.}

\item{folder_path}{Path to the destination folder (i.e. correct: \code{"Folder/Subfolder"}, wrong: \code{"Folder/Subfolder/File.png"}).}

\item{file_name}{File name without file extension (i.e. correct: \code{"File"}, wrong: \code{"File.png"}).}

\item{width_cm}{Plot width in cm.}

\item{height_cm}{Plot height in cm.}

\item{pdf}{Should a pdf file be created and/or immediately opened? Can be either \code{"none"}, \code{"create"} or \code{"open"}.}

\item{png}{Should a png file be created and/or immediately opened? Can be either \code{"none"}, \code{"create"} or \code{"open"}.}

\item{svg}{Should a svg file be created and/or immediately opened? Can be either \code{"none"}, \code{"create"} or \code{"open"}.}

\item{png_dpi}{Plot resolution of png file. Can be a vector of multiple values so that multiple png files will be created.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value from the plot theme.}

\item{png_from_pdf}{If \code{TRUE}, the png file is not exported via \code{ggplot2::ggsave(..., device = "png")}, but instead converted/rendered from the pdf created via \code{ggplot2::ggsave(device = "pdf")}. This can in some cases circumvent issues where pdf and png e.g. have different font sizes.}

\item{svg_device}{If \code{"svg"}, the svg file is not exported via \code{ggplot2::ggsave(..., device = "svg")}, but instead via \code{grDevices::svg()}/\code{grDevices::dev.off()}. This can in some cases circumvent issues with e.g. transparency.}
}
\description{
This function is mostly a wrapper function for \code{ggplot2::ggsave()} that makes it easy to \itemize{ \item{simultaneously export a plot into pdf, svg and/or png (even multiple pngs with different dpi)} \item{immediately open the exported files in your OS}} Additionally, it offers alternatives to rendering via \code{ggplot2::ggsave()} - see arguments \code{png_from_pdf} and \code{svg_device}.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(data = PlantGrowth) +
  aes(y = weight, x = group) +
  geom_point()

gg_export(
  plot_obj = p,
  folder_path = here::here(),
  file_name = "myplot",
  width_cm = 12,
  height_cm = 8,
  pdf = "open",
  png = "open",
  svg = "none",
  png_dpi = c(96, 300)
)
}

}
