% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wb.R
\name{save_wb}
\alias{save_wb}
\title{Save a Workbook object to Excel file}
\usage{
save_wb(
  wb,
  file = here::here("untitled.xlsx"),
  overwrite = FALSE,
  open_file = TRUE,
  ...
)
}
\arguments{
\item{wb}{A Workbook object to write to file}

\item{file}{A character string naming an xlsx file}

\item{overwrite}{If \code{TRUE}, overwrites the previous excel file}

\item{open_file}{If \code{TRUE}, opens the excel file after it was created}

\item{...}{Other arguments passed to \code{openxlsx::saveWorkbook()}}
}
\description{
This function is a wrapper for \code{openxlsx::saveWorkbook()}, but can also open the Excel file immediately after saving.
}
