% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_fit.R
\name{smart_fit}
\alias{smart_fit}
\title{Adjust column widths for flextables}
\usage{
smart_fit(
  ft,
  width = c("A4", "letter", "legal", "executive")[1],
  landscape = FALSE,
  page_margin = "default"
)
}
\arguments{
\item{ft}{Flextable object}

\item{width}{Maximum width allowed for the table. Can be a numeric value (in cm) or one of the following strings referring to standard paper sizes \code{"A4"} (=default), \code{"letter"}, \code{"legal"} or \code{"executive"}.}

\item{landscape}{If \code{TRUE}, landscape width for standard paper size provided in \code{width} is used. Default is \code{FALSE}, i.e. portrait width. Ignored if \code{width} is provided as a numeric value.}

\item{page_margin}{Total page margin (sum of left and right margin) that is subtracted from the total page width of the standard paper width provided in \code{width}. Ignored if \code{width} is provided as a numeric value.}
}
\description{
Adjust column widths for flextables
}
