% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_BioMath.R
\name{theme_BioMath}
\alias{theme_BioMath}
\title{A \{ggplot2\} theme based on the \href{https://www.biomath.de/}{BioMath} corporate design.}
\usage{
theme_BioMath(
  base_size = 11,
  base_family = "",
  base_color = "#001509",
  axis_title_bold = FALSE,
  axis_title_just = "rt",
  ticks = FALSE,
  grid_x = FALSE,
  grid_y = FALSE,
  grid_color = "#C0BCB5",
  grid_linetype = "dotted",
  facette_distance = 1,
  facette_box = FALSE,
  facette_box_color = "#C0BCB5",
  title_size = 14,
  subtitle_size = 10,
  axistext_size = 10,
  ggtext_axis = TRUE,
  whitebg = TRUE,
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family}

\item{base_color}{Base color for text and lines}

\item{axis_title_bold}{Should axis titles be bold?}

\item{axis_title_just}{Justification of axis titles. \code{rt} translates to "right" (x-axis) "top" (y-axis).}

\item{ticks}{Should ticks be drawn on the axes?}

\item{grid_x}{Should there be major grid lines for the x-axis?}

\item{grid_y}{Should there be major grid lines for the y-axis?}

\item{grid_color}{Color of the grid lines}

\item{grid_linetype}{Linetype of the grid lines}

\item{facette_distance}{Distance between facettes}

\item{facette_box}{Should a rectangle be drawn around each facette?}

\item{facette_box_color}{Color of the rectangle around each facette}

\item{title_size}{Title font size, given in pts.}

\item{subtitle_size}{Subtitle font size, given in pts.}

\item{axistext_size}{Axis text font size, given in pts.}

\item{ggtext_axis}{If \code{FALSE}, axis texts are rendered via \code{ggplot2::element_text()} instead of \code{ggtext::element_markdown()}. This can prevent potential rendering issues.}

\item{whitebg}{If \code{TRUE}, plot background is white, otherwise transparent.}

\item{...}{Other arguments passed to the underlying \code{ggplot2::theme_minimal()}}
}
\description{
A \{ggplot2\} theme based on the \href{https://www.biomath.de/}{BioMath} corporate design.
}
\examples{
library(ggplot2)

p1 <- ggplot(data = PlantGrowth) +
  aes(y = weight, x = group, fill = group) +
  geom_bar(stat = "summary", fun = "mean")

BioMathR::palette_getset("BioMath")

p1

p1 + BioMathR::theme_BioMath()
}
